requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
joblib==1.4.2
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
xgboost==2.1.0
